# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OPPO Factory CMMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /production:
    get:
      operationId: getProductionRecords
      tags: [production]
      summary: Get production capacity records
      responses:
        "200":
          description: List of production records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionRecord"
    post:
      operationId: createProductionRecord
      tags: [production]
      summary: Create production record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionRecordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRecord"
  /production/lines:
    get:
      operationId: getProductionLines
      tags: [production]
      summary: Get production lines
      responses:
        "200":
          description: List of production lines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionLine"
    post:
      operationId: createProductionLine
      tags: [production]
      summary: Create production line
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionLineRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionLine"
  /production/lines/{id}:
    put:
      operationId: updateProductionLine
      tags: [production]
      summary: Update production line
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductionLineRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionLine"
...
components:
  schemas:
...
    ProductionLine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        targetCapacityPerHour:
          type: integer
        minimumCapacityPerHour:
          type: integer
        responsibleUserId:
          type: integer
        responsibleUserName:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - targetCapacityPerHour
        - minimumCapacityPerHour
    CreateProductionLineRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetCapacityPerHour:
          type: integer
        minimumCapacityPerHour:
          type: integer
        responsibleUserId:
          type: integer
      required:
        - name
        - targetCapacityPerHour
        - minimumCapacityPerHour
    UpdateProductionLineRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetCapacityPerHour:
          type: integer
        minimumCapacityPerHour:
          type: integer
        responsibleUserId:
          type: integer
        isActive:
          type: boolean
    ProductionRecord:
      type: object
      properties:
        id:
          type: integer
        lineId:
          type: integer
        lineName:
          type: string
        recordedAt:
          type: string
        hour:
          type: integer
        actualCapacity:
          type: integer
        targetCapacity:
          type: integer
        belowLimit:
          type: boolean
        reason:
          type: string
        recordedById:
          type: integer
        recordedByName:
          type: string
        shift:
          type: string
          enum: [morning, afternoon, night]
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - lineId
        - recordedAt
        - actualCapacity
        - recordedById
    CreateProductionRecordRequest:
      type: object
      properties:
        lineId:
          type: integer
        recordedAt:
          type: string
        hour:
          type: integer
        actualCapacity:
          type: integer
        shift:
          type: string
          enum: [morning, afternoon, night]
        reason:
          type: string
        date:
          type: string
      required:
        - lineId
        - actualCapacity
        - shift
        - date
...